Update starfield.js
Key Addition:

Black Center Circle: After drawing all the stars, a black circle is drawn at the center of the canvas. This circle acts as a focal point, giving the illusion that stars are moving away from an infinite space at the center. The radius of the circle is set to 20, but you can adjust this value based on your preference or the size of your canvas to achieve the desired effect.
This modification enhances the visual effect by creating a contrast between the moving stars and the static center, simulating a more immersive warp speed experience.