Update starfield.js
init stars less colour chance 

Color Probability: The probability checks for assigning colors other than white have been significantly reduced. Now, there's only a 0.02% chance for a star to be blue, a cumulative 0.04% chance for a star to be red (including the previous chance for blue), and a cumulative 0.05% chance for a star to be yellow. This means that out of 2000 stars, on average, only 1 star will be blue, 1 additional star might be red, and 1 more might be yellow, making colored stars much rarer.